pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "FileUtils"
version = "0.5.0"
description = "File utilities for data science projects with Azure support"
readme = "README.md"
authors = [{name = "Topi Järvinen"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.3.0",
    "pyyaml>=5.4.1",
    "python-dotenv>=0.19.0",
    "jsonschema>=3.2.0",
]

[tool.setuptools]
packages = ["FileUtils", "FileUtils.core", "FileUtils.storage", "FileUtils.config", "FileUtils.utils"]
package-dir = {"" = "src"}

[project.optional-dependencies]
azure = [
    "azure-storage-blob>=12.0.0",
    "azure-identity>=1.5.0",
]
parquet = ["pyarrow>=7.0.0"]
excel = ["openpyxl>=3.0.9"]
all = [
    "azure-storage-blob>=12.0.0",
    "azure-identity>=1.5.0",
    "pyarrow>=7.0.0",
    "openpyxl>=3.0.9",
]

